refactor out event handling logic
